Updated hand material and response to bullet particle and changed Game Monitor's health variables to ints

+++ Assets/Scripts/GameMonitor.cs
@@ -5,17 +5,17 @@ using UnityEngine;
 public class GameMonitor : MonoBehaviour
 {
     // For Enemy
-    public float rightHandHealth;
+    public int rightHandHealth;
     public AttackPatterns atkPat;
-    public float leftHandHealth;
-    public float headHealth;
-    private float enemyTotalHealth;
+    public int leftHandHealth;
+    public int headHealth;
+    private int enemyTotalHealth;
 
     // For Player
-    public float playerHealth;
+    public int playerHealth;
     
     // For Powerups
-    public float enemyThreshold;
+    public int enemyThreshold;
     private bool powerUp1 = false;
 
     // Update() manages the progress of game in terms of player and enemy health
@@ -24,12 +24,12 @@ public class GameMonitor : MonoBehaviour
         // Update the overall remaining enemy health
         calcEnemyHealth();
         // If player loses all health, player loses
-        if (playerHealth <= 0.0) {
+        if (playerHealth <= 0) {
             Debug.Log("You lose");
             // Enemy Wins
         }
         // If enemy loses all health, player wins
-        else if (enemyTotalHealth <= 0.0) {
+        else if (enemyTotalHealth <= 0) {
             Debug.Log("You Win");
             // Player wins/move on to next phase
         }
@@ -42,10 +42,10 @@ public class GameMonitor : MonoBehaviour
     }
 
     /* Public Functions for player's health to be modified; use these in playerController and adjust if needed*/
-    public void playerTakeDamage(float amt) {
+    public void playerTakeDamage(int amt) {
         playerHealth -= amt;
     }
-    public void playerAddHealth(float amt) {
+    public void playerAddHealth(int amt) {
         playerHealth += amt;
     }
 
@@ -60,10 +60,10 @@ public class GameMonitor : MonoBehaviour
     1 == Left Hand
     2 == Head
     */
-    public void enemyTakeDamage(float amt, int body) {
+    public void enemyTakeDamage(int amt, int body) {
         if (body == 0) {
             rightHandHealth -= amt;
-            if (rightHandHealth <= 0.0) {
+            if (rightHandHealth <= 0) {
                 //Function to disable right hand
                 atkPat.disableRight();
             }
